Add more command regexes, rework existing ones
This was crafted by reading through syncoid v2.2.0 source code and
looking at all possible combinations of options, combining some that
could be combined, and adding new ones.

Added:
  - hold, release (--use-hold)
  - --insecure-direct-connection: refactored mbuffer regex, added
    busybox nc and socat regexes
  - all possible compression commands
  - setting properties on receive: --preserve-properties --preserve-recordsize (untested!)
  - getting syncoid:sync dataset userprop

Reworked:
  - zfs receive
  - getting zfsprops
  - factored out regexes for pipes, redirections, short options,
    mbuffer (due to added --insecure-direct-connection support it has to
    be alternatively busybox nc)

I've tested this with a remote destination, but not with remote source.
All the commands for a remote source (zfs send ...) are thus untested.